Create phase_change
I've been investigating the nature of the phase changes in different dimensions (1 and 3). So far I've found that the 1D case has a phase change at a lower temperature than the 2D case when J = 1. However for larger values of J, there is no phase change! 

I think the next logical step would be to determine the minimum value of J at which there is no phase change. I'll have to update my class to factor the value of J into account